Segmentation fault bugs in disconnection have been fixed.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/trunk/OpenRTM-aist@1351 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/rtm/InPortPushConnector.cpp
@@ -33,7 +33,7 @@ namespace RTC
                                            InPortProvider* provider,
                                            CdrBufferBase* buffer)
     : InPortConnector(profile, buffer),
-      m_provider(provider)
+      m_provider(provider), m_deleteBuffer(buffer == 0 ? true : false)
   {
     rtclog.setLevel("PARANOID");
 
@@ -116,7 +116,7 @@ namespace RTC
     m_provider = 0;
 
     // delete buffer
-    if (m_buffer != 0)
+    if (m_buffer != 0 && m_deleteBuffer == true)
       {
         CdrBufferFactory& bfactory(CdrBufferFactory::instance());
         bfactory.deleteObject(m_buffer);
